requirements.txt
Django==1.8.6
django-registration-redux==1.2
